starting to get if statmments to work.
gonna take a break for the weekend.
they are starting to work. it seems like the best way is to just brute force the stuff and write the code in multiple places. ir would be really complicated.
i feel like my code is starting to become a mess that I dont wanna touch so after i do if statements imma clean it up